+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## NEXT
+* Add support for developing your app using a local Kubernetes cluster.
+  - Includes a default `:development` Kuby environment.
+* Remove rake tasks in favor of a `kuby` executable powered by a vendorized and re-namespaced copy of [GLI](https://github.com/davetron5000/gli).
+  - Executable tries to load as little code as possible, i.e. doesn't load your entire bundle.
+
 ## 0.8.1
 * Fix database config rewriter task.
   - Broke with refactoring of database config code.